-- added data title filter
