+++ src/frontend/src/components/Main/Messages/NewChat.js
@@ -47,6 +47,7 @@ function NewChat (props) {
   const newChatId = useSelector(newChatIdSelector)
   const currentUser = useSelector(currentUserSelector)
   const isChatOpen = useSelector(isTemporaryChatOpenSelector)
+  let dateTitleTemporaryMemory = ''
 
   const currentChat = chatsList.filter(c => c.id === newChat.id)[0]
   const currentChatUsers = currentChat && currentChat.users
@@ -85,6 +86,63 @@ function NewChat (props) {
 
   const getMessageSender = (userId) => {
     return currentChatUsers && currentChatUsers.filter(u => u.id === userId)[0]
+  }
+
+  const checkIfNeedToRenderDateTitle = (time) => {
+    if (dateTitleTemporaryMemory === getDateTitle(time)) {
+      return ''
+    } else {
+      dateTitleTemporaryMemory = getDateTitle(time)
+      return getDateTitle(time)
+    }
+  }
+
+  const getDateTitle = (time) => {
+    const localTime = new Date()
+    switch (true) {
+      case localTime.getFullYear() === +time.split('T')[0].split('-')[0] && +time.split('T')[0].split('-')[2] !== localTime.getDate():
+        return time.split('T')[1].split('.')[2] + '.' + time.split('T')[1].split('.')[1]
+      case localTime.getDate() !== +time.split('T')[0].split('-')[2] && localTime.getFullYear() !== +time.split('T')[0].split('-')[0]:
+        return time.split('T')[0].split('-')[2] + ' ' + getMonthText(time.split('T')[0].split('-')[1]) + ' ' + time.split('T')[0].split('-')[0]
+      default:
+        return time.split('T')[0].split('-')[2] + ' ' + getMonthText(time.split('T')[0].split('-')[1])
+    }
+  }
+
+  const getDate = (time) => {
+    const splitDate = time.split('T')[1].split('.')[0].split(':').slice(0, 2)
+    return splitDate[0] + ':' + splitDate[1]
+  }
+
+  const getMonthText = (date) => {
+    switch (date) {
+      case '01':
+        return 'JANUARY'
+      case '02':
+        return 'FEBRUARY'
+      case '03':
+        return 'MARCH'
+      case '04':
+        return 'APRIL'
+      case '05':
+        return 'MAY'
+      case '06':
+        return 'JUNE'
+      case '07':
+        return 'JULY'
+      case '08':
+        return 'AUGUST'
+      case '09':
+        return 'SEPTEMBER'
+      case '10':
+        return 'OCTOBER'
+      case '11':
+        return 'NOVEMBER'
+      case '12':
+        return 'DECEMBER'
+      default:
+        return 'incorrect data'
+    }
   }
 
   return (
@@ -143,7 +201,8 @@ function NewChat (props) {
                 {userChatMessages[newChat.id] && userChatMessages[newChat.id].map(m => <UserMessage key={m.id}
                   messageSender={getMessageSender(m.userId)}
                   text={m.text}
-                  time={m.createdDate}/>)}
+                  timeTitle={checkIfNeedToRenderDateTitle(m.createdDate)}
+                  timeSent={getDate(m.createdDate)}/>)}
               </li>
             </ul>
           </div>
